requirements.txt
wget
biopython
pandas
mysql-connector
